refactor: extract EditSpawnerTrait
Extracts EditSpawnerTrait, which defines:

- `getEditor() : Editor`
- `createTempFileWithContents(string $contents) : string`
- `unlinkTempFile(string $filename) : void`
- `public ?Editor $editor`
- `public ?string $mockTempFile`

The trait should be composed into any class that would normally call
`new Editor()`.

Extracted from EditChangelogLinksListener, and added to each of
EditChangelogVersionListener and AbstractEditConfigListener.
