+++ src/Changelog/EditChangelogLinksListener.php
@@ -11,19 +11,14 @@ namespace Phly\KeepAChangelog\Changelog;
 
 use Phly\KeepAChangelog\Common\ChangelogEditor;
 use Phly\KeepAChangelog\Common\ChangelogEntry;
-use Phly\KeepAChangelog\Common\Editor;
+use Phly\KeepAChangelog\Common\EditSpawnerTrait;
 
-use function file_exists;
 use function file_get_contents;
-use function file_put_contents;
-use function is_string;
-use function sprintf;
-use function sys_get_temp_dir;
-use function uniqid;
-use function unlink;
 
 class EditChangelogLinksListener
 {
+    use EditSpawnerTrait;
+
     public function __invoke(EditChangelogLinksEvent $event) : void
     {
         $changelog = $event->config()->changelogFile();
@@ -55,35 +50,6 @@ class EditChangelogLinksListener
         $event->editComplete($changelog);
     }
 
-    /**
-     * Creates a temporary file with the changelog contents.
-     */
-    private function createTempFileWithContents(string $contents) : string
-    {
-        if (is_string($this->mockTempFile)) {
-            return $this->mockTempFile;
-        }
-
-        $filename = sprintf('%s.md', uniqid('KAC', true));
-        $path     = sprintf('%s/%s', sys_get_temp_dir(), $filename);
-
-        file_put_contents($path, $contents);
-
-        return $path;
-    }
-
-    /**
-     * Removes the temp file generated by createTempFileWithContents
-     */
-    private function unlinkTempFile(string $filename) : void
-    {
-        if (is_string($this->mockTempFile) || ! file_exists($filename)) {
-            return;
-        }
-
-        unlink($filename);
-    }
-
     private function getChangelogEditor() : ChangelogEditor
     {
         if ($this->changelogEditor instanceof ChangelogEditor) {
@@ -91,15 +57,6 @@ class EditChangelogLinksListener
         }
 
         return new ChangelogEditor();
-    }
-
-    private function getEditor() : Editor
-    {
-        if ($this->editor instanceof Editor) {
-            return $this->editor;
-        }
-
-        return new Editor();
     }
 
     /**
@@ -111,24 +68,4 @@ class EditChangelogLinksListener
      * @var null|ChangelogEditor
      */
     public $changelogEditor;
-
-    /**
-     * Provide an Editor instance to use.
-     *
-     * For testing purposes only.
-     *
-     * @internal
-     * @var null|Editor
-     */
-    public $editor;
-
-    /**
-     * Provide a mock temporary filename.
-     *
-     * For testing purposes only.
-     *
-     * @internal
-     * @var null|string
-     */
-    public $mockTempFile;
 }
