+++ src/Version/EditChangelogVersionListener.php
@@ -10,16 +10,14 @@ declare(strict_types=1);
 namespace Phly\KeepAChangelog\Version;
 
 use Phly\KeepAChangelog\Common\ChangelogEditor;
-use Phly\KeepAChangelog\Common\Editor;
+use Phly\KeepAChangelog\Common\EditSpawnerTrait;
 
 use function file_get_contents;
-use function file_put_contents;
-use function sprintf;
-use function sys_get_temp_dir;
-use function uniqid;
 
 class EditChangelogVersionListener
 {
+    use EditSpawnerTrait;
+
     public function __invoke(EditChangelogVersionEvent $event) : void
     {
         $changelogEntry = $event->changelogEntry();
@@ -27,7 +25,7 @@ class EditChangelogVersionListener
             $changelogEntry->contents
         );
 
-        $status = (new Editor())->spawnEditor(
+        $status = $this->getEditor()->spawnEditor(
             $event->output(),
             $event->editor(),
             $tempFile
@@ -45,16 +43,5 @@ class EditChangelogVersionListener
         );
 
         $event->editComplete();
-    }
-
-    /**
-     * Creates a temporary file with the changelog contents.
-     */
-    private function createTempFileWithContents(string $contents) : string
-    {
-        $filename = sprintf('%s.md', uniqid('KAC', true));
-        $path     = sprintf('%s/%s', sys_get_temp_dir(), $filename);
-        file_put_contents($path, $contents);
-        return $path;
     }
 }
